6. REALTIME_PROJECTS/9. generic_ingest_search_app/backend/pyproject.toml
[project]
name = "generic-ingest-search-api"
version = "1.0.0"
description = "Full-stack OpenSearch ingestion and search API"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "python-multipart==0.0.17",
    "pydantic==2.10.3",
    "pydantic-settings==2.6.1",
    "opensearch-py==3.0.0",
    "opensearch-py-ml==1.3.0",
    "pandas==2.2.3",
    "openpyxl==3.1.5",
    "redis==5.2.0",
    "python-jose[cryptography]==3.3.0",
    "pyjwt==2.10.1",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest==8.3.4",
    "httpx==0.28.1",
]
